Moved fsm_msgTransactionSign to fsm_skycoin.c
