+++ tiny-firmware/firmware/fsm_skycoin_impl.c
@@ -154,3 +154,127 @@ ErrCode_t msgSkycoinAddressImpl(SkycoinAddress *msg, ResponseSkycoinAddress *res
     }
     return ErrOk;
 }
+
+ErrCode_t
+msgTransactionSignImpl(TransactionSign *msg, ErrCode_t (*funcConfirmTxn)(char *, char *, TransactionSign *, uint32_t),
+                       ResponseTransactionSign *resp) {
+    if (msg->nbIn > sizeof(msg->transactionIn) / sizeof(*msg->transactionIn)) {
+        return ErrInvalidArg;
+    }
+    if (msg->nbOut > sizeof(msg->transactionOut) / sizeof(*msg->transactionOut)) {
+        return ErrInvalidArg;
+    }
+#if EMULATOR
+    printf("%s: %d. nbOut: %d\n",
+        _("Transaction signed nbIn"),
+        msg->nbIn, msg->nbOut);
+
+    for (uint32_t i = 0; i < msg->nbIn; ++i) {
+        printf("Input: addressIn: %s, index: %d\n",
+            msg->transactionIn[i].hashIn, msg->transactionIn[i].index);
+    }
+    for (uint32_t i = 0; i < msg->nbOut; ++i) {
+        printf("Output: coin: %" PRIu64 ", hour: %" PRIu64 " address: %s address_index: %d\n",
+            msg->transactionOut[i].coin, msg->transactionOut[i].hour,
+            msg->transactionOut[i].address, msg->transactionOut[i].address_index);
+    }
+#endif
+    Transaction transaction;
+    transaction_initZeroTransaction(&transaction);
+    for (uint32_t i = 0; i < msg->nbIn; ++i) {
+        uint8_t hashIn[32];
+        writebuf_fromhexstr(msg->transactionIn[i].hashIn, hashIn);
+        transaction_addInput(&transaction, hashIn);
+    }
+    for (uint32_t i = 0; i < msg->nbOut; ++i) {
+        char strHour[30];
+        char strCoin[30];
+        char strValue[20];
+        char *coinString = msg->transactionOut[i].coin == 1000000 ? _("coin") : _("coins");
+        char *hourString = (msg->transactionOut[i].hour == 1 || msg->transactionOut[i].hour == 0) ? _("hour") : _(
+                "hours");
+        char *strValueMsg = sprint_coins(msg->transactionOut[i].coin, SKYPARAM_DROPLET_PRECISION_EXP, sizeof(strValue),
+                                         strValue);
+        if (strValueMsg == NULL) {
+            // FIXME: For Skycoin coin supply and precision buffer size should be enough
+            strcpy(strCoin, "too many coins");
+        }
+        sprintf(strCoin, "%s %s %s", _("send"), strValueMsg, coinString);
+        sprintf(strHour, "%"
+        PRIu64
+        " %s", msg->transactionOut[i].hour, hourString);
+
+        if (msg->transactionOut[i].has_address_index) {
+            uint8_t pubkey[33] = {0};
+            uint8_t seckey[32] = {0};
+            size_t size_address = 36;
+            char address[36] = {0};
+            ErrCode_t ret = fsm_getKeyPairAtIndex(1, pubkey, seckey, NULL, msg->transactionOut[i].address_index);
+            if (ret != ErrOk) {
+                return ret;
+            }
+            if (!skycoin_address_from_pubkey(pubkey, address, &size_address)) {
+                return ErrAddressGeneration;
+            }
+            if (strcmp(msg->transactionOut[i].address, address) != 0) {
+                // fsm_sendFailure(FailureType_Failure_AddressGeneration, _("Wrong return address"));
+#if EMULATOR
+                printf("Internal address: %s, message address: %s\n", address, msg->transactionOut[i].address);
+                printf("Comparaison size %ld\n", size_address);
+#endif
+                return ErrAddressGeneration;
+            }
+        } else {
+            // NOTICE: A single output per address is assumed
+            ErrCode_t err = funcConfirmTxn(strCoin, strHour, msg, i);
+            if (err != ErrOk)
+                return err;
+        }
+        transaction_addOutput(&transaction, msg->transactionOut[i].coin, msg->transactionOut[i].hour,
+                              msg->transactionOut[i].address);
+    }
+
+    if (!protectPin(false)) {
+        return ErrPinRequired;
+    }
+
+    for (uint32_t i = 0; i < msg->nbIn; ++i) {
+        uint8_t digest[32] = {0};
+        transaction_msgToSign(&transaction, i, digest);
+        // Only sign inputs owned by Skywallet device
+        if (msg->transactionIn[i].has_index) {
+            if (msgSignTransactionMessageImpl(digest, msg->transactionIn[i].index,
+                                              resp->signatures[resp->signatures_count]) != ErrOk) {
+                //fsm_sendFailure(FailureType_Failure_InvalidSignature, NULL);
+                //layoutHome();
+                return ErrInvalidSignature;
+            }
+        } else {
+            // Null sig
+            uint8_t signature[65];
+            memset(signature, 0, sizeof(signature));
+            tohex(resp->signatures[resp->signatures_count], signature, sizeof(signature));
+        }
+        resp->signatures_count++;
+#if EMULATOR
+        char str[64];
+        tohex(str, (uint8_t*)digest, 32);
+        printf("Signing message:  %s\n", str);
+        printf("Signed message:  %s\n", resp->signatures[i]);
+        printf("Nb signatures: %d\n", resp->signatures_count);
+#endif
+    }
+    if (resp->signatures_count != msg->nbIn) {
+        // Ensure number of sigs and inputs is the same. Mismatch should never happen.
+        return ErrFailed;
+    }
+#if EMULATOR
+    char str[64];
+    tohex(str, transaction.innerHash, 32);
+    printf("InnerHash %s\n", str);
+    printf("Signed message:  %s\n", resp->signatures[0]);
+    printf("Nb signatures: %d\n", resp->signatures_count);
+#endif
+    //layoutHome();
+    return ErrOk;
+}
